Some updates, added index_no_index functionality, deleted the abstract databse class removed,updated our drop table if exists functionality, moved over getrandomstring function to MySQLDatabase, and added new attribute for MySQLDatabase to differentiate between mysql and sqlite
